Upgrade xil_cloud to basic_lane_follow scenario in CDASim (#415)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
This PR updates the xil_cloud to xil_basic_lanefollow_scenario, which it
really was during K900 testing.
Any of the ns3, infrastructure, traffic control etc was not really used
in the scenario, so I removed them.
The reason is that we have other scenarios that meaningfully test above
functionality.

Therefore this scenario is currently obselete and not purposeful on its
own that we can assign a purpose here.
The purpose is to just test basic lanefollow functionality before the
developer go to a real world.
This was identified useful for smoke testing Humble upgrades for example
that will help us isolate components of carma-platform and ros1_bridge
to indentify business logic issues early on. This was suggested by
Andrew Loughran who is co-leading the regression testing effort with
Kevin.

Document:
https://usdot-carma.atlassian.net/wiki/spaces/CRMPLT/pages/3335749682/CARMA+Platform+C-ADS+smoke+testing+procedure

<!--- Describe your changes in detail -->

## Related Issue
NA
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
Humble Upgrade Smoke testing plan creation
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
TODO: Simulation PC 2
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [X] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [X] I have read the **CONTRIBUTING** document.
[CARMA Contributing
Guide](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
- [X] I have added tests to cover my changes.
- [X] All new and existing tests passed.