Add global parameter override yaml file (#475)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
Adding a global parameter override yaml file to carma-config. This will
help version control any map or scenario specific parameters using
carma-config. GlobalParamsOverride.yaml file will include the parameters
that will override the params in yaml file that is included the
individual packages.

This also has a benefit of not needing to create docker volumes to
multiple packages. We can add the desired parameter in this file to
override the package level parameter.

Design document:
https://usdot-carma.atlassian.net/wiki/spaces/CRMPLT/pages/3759669249/High+Level+Design+-+CARMA+Config+Update+Configuration+Management
<!--- Describe your changes in detail -->

## Related Issue
CDAS-74
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
- Some tests and demos require special parameters in carma-platform
docker image. Those values are not released so it takes time to
re-discover those again. With the new way, carma-config truly houses all
the configurations needed to make the demo work again (TODO: in the near
future, we are planning to include necessary route and map files to
fully complete the configurations as well, so map and routes version
controlled as well and works out of the box).
- CDASim had multiple params.yaml file for each plugins. This approach
gives them the unified way to change the parameters.
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Locally on sim PC. All nested namespaces were tested.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [X] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [X] I have read the **CONTRIBUTING** document.
[CARMA Contributing
Guide](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
- [X] I have added tests to cover my changes.
- [X] All new and existing tests passed.